Start refactoring update functions
