+++ mapboard/topology_manager/__init__.py
@@ -1,5 +1,7 @@
 from .manager import TopologyManager
 from .config import create_context
 
-__all__ = ["TopologyManager", "create_context"]
+from .commands import update
 
+
+__all__ = ["TopologyManager", "create_context", "update"]
