+++ pyproject.toml
@@ -26,6 +26,7 @@ dependencies = [
     "pytest-profiling>=1.7.0,<2",
     "addict>=2.4.0,<3",
     "geoalchemy2>=0.20.0",
+    "pydantic>=2.13.4",
 ]
 
 [project.scripts]
